DRILL-406: Move JSON integer range to use BigInt instead of Int

+++ exec/java-exec/src/main/java/org/apache/drill/exec/schema/json/jackson/JacksonHelper.java
@@ -33,8 +33,8 @@ public class JacksonHelper {
   public static final MajorType REPEATED_BOOLEAN_TYPE = Types.repeated(MinorType.BIT);
   public static final MajorType ARRAY_TYPE = Types.repeated(MinorType.LATE);
   public static final MajorType MAP_TYPE = Types.repeated(MinorType.MAP);
-  public static final MajorType INT_TYPE = Types.optional(MinorType.INT);
-  public static final MajorType REPEATED_INT_TYPE = Types.repeated(MinorType.INT);
+  public static final MajorType INT_TYPE = Types.optional(MinorType.BIGINT);
+  public static final MajorType REPEATED_INT_TYPE = Types.repeated(MinorType.BIGINT);
   public static final MajorType FLOAT_TYPE = Types.optional(MinorType.FLOAT4);
   public static final MajorType REPEATED_FLOAT_TYPE = Types.repeated(MinorType.FLOAT4);
   public static final MajorType NULL_TYPE = Types.optional(MinorType.LATE);
@@ -82,8 +82,8 @@ public class JacksonHelper {
 
   public static Object getValueFromFieldType(JsonParser parser, MinorType fieldType) throws IOException {
     switch (fieldType) {
-      case INT:
-        return parser.getIntValue();
+      case BIGINT:
+        return parser.getLongValue();
       case VARCHAR:
         return parser.getValueAsString();
       case FLOAT4:
